♻️ Create markup/index.tsx
